Add MDN WebDocs
Web Docs is a great resource for developers with articles, guides and tutorials in addition to being a well maintained reference guide. It would be best used in conjunction with some of the previously listed resources such as FreeCodeCamp and The Odin Project to provide more background information for topics covered in those courses.